+++ packages/core/src/bin/utils/run.ts
@@ -16,6 +16,7 @@ import {
 } from "@/utils/checkpoint.js";
 import { chunk } from "@/utils/chunk.js";
 import { formatEta, formatPercentage } from "@/utils/format.js";
+import { recordAsyncGenerator } from "@/utils/generators.js";
 import { createMutex } from "@/utils/mutex.js";
 import { never } from "@/utils/never.js";
 import { createRequestQueue } from "@/utils/requestQueue.js";
@@ -147,7 +148,23 @@ export async function run({
   }
 
   // Run historical indexing until complete.
-  for await (const events of sync.getEvents()) {
+  for await (const events of recordAsyncGenerator(
+    sync.getEvents(),
+    (params) => {
+      common.metrics.ponder_historical_phase_duration.inc(
+        { phase: "extract" },
+        params.await,
+      );
+      common.metrics.ponder_historical_phase_duration.inc(
+        { phase: "transform" },
+        params.yield,
+      );
+      common.metrics.ponder_historical_phase_duration.inc(
+        { phase: "total" },
+        params.total,
+      );
+    },
+  )) {
     if (events.length > 0) {
       await database.retry(async () => {
         await database
@@ -180,7 +197,7 @@ export async function run({
                 common.metrics.ponder_historical_completed_indexing_seconds.set(
                   { network: network.name },
                   Math.max(
-                    checkpoint.blockTimestamp -
+                    Number(checkpoint.blockTimestamp) -
                       sync.seconds[network.name]!.start -
                       sync.seconds[network.name]!.cached,
                     0,
@@ -188,7 +205,7 @@ export async function run({
                 );
                 common.metrics.ponder_indexing_timestamp.set(
                   { network: network.name },
-                  checkpoint.blockTimestamp,
+                  Number(checkpoint.blockTimestamp),
                 );
               }
 
@@ -356,13 +373,13 @@ export async function run({
 
               common.metrics.ponder_indexing_timestamp.set(
                 { network: network.name },
-                decodeCheckpoint(checkpoint).blockTimestamp,
+                Number(decodeCheckpoint(checkpoint).blockTimestamp),
               );
             } else {
               for (const network of indexingBuild.networks) {
                 common.metrics.ponder_indexing_timestamp.set(
                   { network: network.name },
-                  decodeCheckpoint(checkpoint).blockTimestamp,
+                  Number(decodeCheckpoint(checkpoint).blockTimestamp),
                 );
               }
             }