+++ packages/core/src/sync-realtime/index.ts
@@ -479,9 +479,10 @@ export const createRealtimeSync = (
 
         const msg = `Encountered unrecoverable '${args.network.name}' reorg beyond finalized block ${hexToNumber(finalizedBlock.number)}`;
 
-        args.common.logger.warn({ service: "realtime", msg });
-
-        throw new Error(msg);
+        const error = new Error(msg);
+        args.common.logger.error({ service: "realtime", msg });
+        args.onFatalError(error);
+        return;
       } else {
         remoteBlock = await _eth_getBlockByHash(args.requestQueue, {
           hash: remoteBlock.parentHash,